refactored code into monolithic application
