+++ .gitignore
@@ -1,1 +1,3 @@
 credentials.json
+.pytest_cache
+__pycache__
